feat: cleanup and wip work
